+++ src/main/java/com/kosa/backend/payment/repository/CouponRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface CouponRepository extends JpaRepository<Coupon, Integer> {
     // JpaRepository provides basic CRUD method.
-    List<Coupon> findByUser_Id(Long userId);  // 특정 유저의 쿠폰 목록을 조회하는 메서드
+    List<Coupon> findByUser_Id(int userId);  // 특정 유저의 쿠폰 목록을 조회하는 메서드
 }
